refactor: deduplicate max effective square size
