+++ x/blob/keeper/params.go
@@ -24,7 +24,11 @@ func (k Keeper) GasPerBlobByte(ctx sdk.Context) (res uint32) {
 	return res
 }
 
-// GovMaxSquareSize returns the GovMaxSquareSize param
+// GovMaxSquareSize returns the GovMaxSquareSize param.
+//
+// Note: it is unlikely that you want to use this value directly because
+// governance can modify it to be anything. Most use-cases will want to use
+// squaresize.MaxEffective instead of GovMaxSquareSize.
 func (k Keeper) GovMaxSquareSize(ctx sdk.Context) (res uint64) {
 	k.paramStore.Get(ctx, types.KeyGovMaxSquareSize, &res)
 	return res
